Ej4C

+++ src/main/java/com/amalagonj/ej2notes/controllers/NoteController.java
@@ -1,7 +1,10 @@
 package com.amalagonj.ej2notes.controllers;
+import com.amalagonj.ej2notes.exception.ConcurrencyConflictException;
 import com.amalagonj.ej2notes.exception.NoteNotFoundException;
 import com.amalagonj.ej2notes.model.Note;
 import com.amalagonj.ej2notes.repository.NoteRepository;
+import jakarta.validation.Valid;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -36,7 +39,27 @@ public class NoteController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteNoteById(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteNoteById(@PathVariable Long id) {
+        if (!noteRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         noteRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PutMapping("/{id}")
+    public Note updateNote(@PathVariable Long id, @Valid @RequestBody Note noteDetails) {
+
+        Note note = noteRepository.findById(id)
+                .orElseThrow(() -> new NoteNotFoundException(id)); // Reutiliza la excepción del Nivel 1
+
+        if ("CONFLICTO".equalsIgnoreCase(noteDetails.getContent())) {
+            throw new ConcurrencyConflictException("Nota ID " + id);
+        }
+
+        note.setTitle(noteDetails.getTitle());
+        note.setContent(noteDetails.getContent());
+
+        return noteRepository.save(note);
     }
 }
